Fixed formatting to make CI tests pass
