+++ CONTRIBUTORS.txt
@@ -247,3 +247,5 @@ contributors:
 * Lucas Cimon: contributor
 
 * Mike Miller: contributor
+
+* Siva Mahadevan: contributor
