+++ pylint/pyreverse/__init__.py
@@ -1,4 +1,3 @@
-
 # Licensed under the GPL: https://www.gnu.org/licenses/old-licenses/gpl-2.0.html
 # For details: https://github.com/PyCQA/pylint/blob/master/COPYING
