+++ pylint/test/conftest.py
@@ -8,9 +8,11 @@ from pylint.lint import PyLinter
 # pylint: disable=no-name-in-module
 from pylint.testutils import MinimalTestReporter
 
+
 def pytest_addoption(parser):
-    parser.addoption('--spelling-dict-name', action='store')
-    parser.addoption('--spelling-dict-paths', action='store')
+    parser.addoption("--spelling-dict-name", action="store")
+    parser.addoption("--spelling-dict-paths", action="store")
+
 
 @pytest.fixture
 def linter(checker, register, enable, disable, reporter):
