+++ tox.ini
@@ -6,8 +6,9 @@ skip_missing_interpreters = true
 deps =
    git+https://github.com/pycqa/astroid@master
    isort
-
-commands = pylint -rn --rcfile={toxinidir}/pylintrc --load-plugins=pylint.extensions.docparams, pylint.extensions.mccabe {envsitepackagesdir}/pylint
+; McCabe complexity checking is currently disabled in CI, too many warnings. Uncomment this line if desired.
+; commands = pylint -rn --rcfile={toxinidir}/pylintrc --load-plugins=pylint.extensions.docparams,pylint.extensions.mccabe {envsitepackagesdir}/pylint
+commands = pylint -rn --rcfile={toxinidir}/pylintrc --load-plugins=pylint.extensions.docparams {envsitepackagesdir}/pylint
 
 
 [testenv:formatting]
@@ -37,7 +38,7 @@ setenv =
     COVERAGE_FILE = {toxinidir}/.coverage.{envname}
 
 commands =
-    python -Wi {envsitepackagesdir}/coverage run -m pytest {envsitepackagesdir}/pylint/test/ {posargs:}
+    python -Wi {envsitepackagesdir}/coverage run -m pytest {posargs:} {envsitepackagesdir}/pylint/test/
 
     ; Transform absolute path to relative path
     ; for compatibility with coveralls.io and fix 'source not available' error.
@@ -54,7 +55,7 @@ deps =
    hunspell
 
 commands =
-    python -Wi -m pytest {envsitepackagesdir}/pylint/test/ -k unittest_checker_spelling {posargs:} 
+    python -Wi -m pytest {posargs:} {envsitepackagesdir}/pylint/test/unittest_checker_spelling.py
 
 changedir = {toxworkdir}
